# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Primeagent
  version: 1.7.0
paths:
...
  /api/v1/validate/code:
    post:
      tags:
      - Validate
      summary: Post Validate Code
      operationId: post_validate_code_api_v1_validate_code_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Code'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodeValidationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
  /api/v1/validate/prompt:
    post:
      tags:
      - Validate
      summary: Post Validate Prompt
      operationId: post_validate_prompt_api_v1_validate_prompt_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidatePromptRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromptValidationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/users/:
    post:
      tags:
      - Users
      summary: Add User
      description: Add a new user to the database.
      operationId: add_user_api_v1_users__post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Users
      summary: Read All Users
      description: Retrieve a list of users from the database with pagination.
      operationId: read_all_users_api_v1_users__get
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
      parameters:
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 10
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/users/whoami:
    get:
      tags:
      - Users
      summary: Read Current User
      description: Retrieve the current user's data.
      operationId: read_current_user_api_v1_users_whoami_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
  /api/v1/users/{user_id}:
    patch:
      tags:
      - Users
      summary: Patch User
      description: Update an existing user's data.
      operationId: patch_user_api_v1_users__user_id__patch
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Users
      summary: Delete User
      description: Delete a user from the database.
      operationId: delete_user_api_v1_users__user_id__delete
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Delete User Api V1 Users  User Id  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/users/{user_id}/reset-password:
    patch:
      tags:
      - Users
      summary: Reset Password
      description: Reset a user's password.
      operationId: reset_password_api_v1_users__user_id__reset_password_patch
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Code:
      properties:
        code:
          type: string
          title: Code
      type: object
      required:
      - code
      title: Code
...
    CodeValidationResponse:
      properties:
        imports:
          type: object
          title: Imports
        function:
          type: object
          title: Function
      type: object
      required:
      - imports
      - function
      title: CodeValidationResponse
...
    FrontendNodeRequest-Input:
      properties:
        template:
          type: object
          title: Template
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        icon:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon
        is_input:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Input
        is_output:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Output
        is_composition:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Composition
        base_classes:
          items:
            type: string
          type: array
          title: Base Classes
        name:
          type: string
          title: Name
          default: ''
        display_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Display Name
          default: ''
        priority:
          anyOf:
          - type: integer
          - type: 'null'
          title: Priority
        documentation:
          type: string
          title: Documentation
          default: ''
        minimized:
          type: boolean
          title: Minimized
          default: false
        custom_fields:
          anyOf:
          - type: object
          - type: 'null'
          title: Custom Fields
        output_types:
          items:
            type: string
          type: array
          title: Output Types
          default: []
        full_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Path
        pinned:
          type: boolean
          title: Pinned
          default: false
        conditional_paths:
          items:
            type: string
          type: array
          title: Conditional Paths
          default: []
        frozen:
          type: boolean
          title: Frozen
          default: false
        outputs:
          items:
            $ref: '#/components/schemas/Output'
          type: array
          title: Outputs
          default: []
        field_order:
          items:
            type: string
          type: array
          title: Field Order
          default: []
        beta:
          type: boolean
          title: Beta
          default: false
        legacy:
          type: boolean
          title: Legacy
          default: false
        replacement:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Replacement
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
        edited:
          type: boolean
          title: Edited
          default: false
        metadata:
          type: object
          title: Metadata
          default: {}
        tool_mode:
          type: boolean
          title: Tool Mode
          default: false
      type: object
      required:
      - template
      - base_classes
      title: FrontendNodeRequest
    FrontendNodeRequest-Output:
      properties:
        template:
          type: object
          title: Template
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        icon:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon
        is_input:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Input
        is_output:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Output
        is_composition:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Composition
        base_classes:
          items:
            type: string
          type: array
          title: Base Classes
        name:
          type: string
          title: Name
          default: ''
        display_name:
          type: string
          title: Display Name
        priority:
          anyOf:
          - type: integer
          - type: 'null'
          title: Priority
        documentation:
          type: string
          title: Documentation
          default: ''
        minimized:
          type: boolean
          title: Minimized
          default: false
        custom_fields:
          anyOf:
          - type: object
          - type: 'null'
          title: Custom Fields
        output_types:
          items:
            type: string
          type: array
          title: Output Types
          default: []
        full_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Path
        pinned:
          type: boolean
          title: Pinned
          default: false
        conditional_paths:
          items:
            type: string
          type: array
          title: Conditional Paths
          default: []
        frozen:
          type: boolean
          title: Frozen
          default: false
        outputs:
          items:
            $ref: '#/components/schemas/Output'
          type: array
          title: Outputs
          default: []
        field_order:
          items:
            type: string
          type: array
          title: Field Order
          default: []
        beta:
          type: boolean
          title: Beta
          default: false
        legacy:
          type: boolean
          title: Legacy
          default: false
        replacement:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Replacement
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
        edited:
          type: boolean
          title: Edited
          default: false
        metadata:
          type: object
          title: Metadata
          default: {}
        tool_mode:
          type: boolean
          title: Tool Mode
          default: false
      type: object
      required:
      - template
      - base_classes
      title: FrontendNodeRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Output:
      properties:
        types:
          items:
            type: string
          type: array
          title: Types
          default: []
        selected:
          anyOf:
          - type: string
          - type: 'null'
          title: Selected
        name:
          type: string
          title: Name
          description: The name of the field.
        hidden:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Hidden
        display_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Display Name
        method:
          anyOf:
          - type: string
          - type: 'null'
          title: Method
        value:
          anyOf:
          - {}
          - type: 'null'
          title: Value
          default: __UNDEFINED__
        cache:
          type: boolean
          title: Cache
          default: true
        required_inputs:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Required Inputs
        allows_loop:
          type: boolean
          title: Allows Loop
          default: false
        group_outputs:
          type: boolean
          title: Group Outputs
          default: false
        options:
          anyOf:
          - $ref: '#/components/schemas/OutputOptions'
          - type: 'null'
        tool_mode:
          type: boolean
          title: Tool Mode
          default: true
      type: object
      required:
      - name
      title: Output
    OutputOptions:
      properties:
        filter:
          anyOf:
          - type: string
          - type: 'null'
          title: Filter
      type: object
      title: OutputOptions
...
    PromptValidationResponse:
      properties:
        input_variables:
          items: {}
          type: array
          title: Input Variables
        frontend_node:
          anyOf:
          - $ref: '#/components/schemas/FrontendNodeRequest-Output'
          - type: 'null'
      type: object
      required:
      - input_variables
      title: PromptValidationResponse
...
    UserCreate:
      properties:
        username:
          type: string
          title: Username
        password:
          type: string
          title: Password
        optins:
          anyOf:
          - type: object
          - type: 'null'
          title: Optins
          default:
            github_starred: false
            dialog_dismissed: false
            discord_clicked: false
      type: object
      required:
      - username
      - password
      title: UserCreate
    UserRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        username:
          type: string
          title: Username
        profile_image:
          anyOf:
          - type: string
          - type: 'null'
          title: Profile Image
        store_api_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Store Api Key
        is_active:
          type: boolean
          title: Is Active
        is_superuser:
          type: boolean
          title: Is Superuser
        create_at:
          type: string
          format: date-time
          title: Create At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        last_login_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Login At
        optins:
          anyOf:
          - type: object
          - type: 'null'
          title: Optins
      type: object
      required:
      - username
      - profile_image
      - store_api_key
      - is_active
      - is_superuser
      - create_at
      - updated_at
      - last_login_at
      title: UserRead
    UserUpdate:
      properties:
        username:
          anyOf:
          - type: string
          - type: 'null'
          title: Username
        profile_image:
          anyOf:
          - type: string
          - type: 'null'
          title: Profile Image
        password:
          anyOf:
          - type: string
          - type: 'null'
          title: Password
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
        is_superuser:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Superuser
        last_login_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Login At
        optins:
          anyOf:
          - type: object
          - type: 'null'
          title: Optins
      type: object
      title: UserUpdate
...
    UsersResponse:
      properties:
        total_count:
          type: integer
          title: Total Count
        users:
          items:
            $ref: '#/components/schemas/UserRead'
          type: array
          title: Users
      type: object
      required:
      - total_count
      - users
      title: UsersResponse
    ValidatePromptRequest:
      properties:
        name:
          type: string
          title: Name
        template:
          type: string
          title: Template
        custom_fields:
          anyOf:
          - type: object
          - type: 'null'
          title: Custom Fields
        frontend_node:
          anyOf:
          - $ref: '#/components/schemas/FrontendNodeRequest-Input'
          - type: 'null'
      type: object
      required:
      - name
      - template
      title: ValidatePromptRequest
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: api/v1/login
    API key query:
      type: apiKey
      in: query
      name: x-api-key
    API key header:
      type: apiKey
      in: header
      name: x-api-key